Fix styles uploads
Forgot to move uploads logic to normalized styles
